tech(pat-querystring): lazify pattern, split registration from implementation
pat-querystring contributed ~33KB src to the eager patterns chunk on every
page, though the widget only appears on Collection edit forms and the
querystring criteria UI. It also statically imported the contentbrowser
pattern, dragging an eager edge into pat-contentbrowser.

Split it into a thin registration module and querystring--implementation.js
(holding the widget and the contentbrowser import). The implementation is
grafted onto the single pattern instance in init(), so external consumers
that reach into instance state (pat-structure's textfilter reads
this.queryString.$sortOn/$sortOrder) keep working. Public options and
emitted markup are unchanged.

The implementation and its contentbrowser edge now load only on pages with a
.pat-querystring element.
